% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSfateSignificanceScore.R
\name{get_SignificanceScore}
\alias{get_SignificanceScore}
\title{calculate the Significance Score for candidate TF pairs}
\usage{
get_SignificanceScore(scExp, tf_1, tf_2, entropy_correction_para)
}
\arguments{
\item{correction}{parameters to prevent the occurrence of zero entropy due to all computational data being zero, resulting in NA (Not Available) values}
}
\value{
A list of gene pair, ranked by the significance score.
}
\description{
function to fit time series data by nls
}
