% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSfateTF.R
\name{get_singleTF_BSfate_rank}
\alias{get_singleTF_BSfate_rank}
\title{get_single_BSfate_rank}
\usage{
get_singleTF_BSfate_rank(SignificanceScore)
}
\value{
A list of TF, ranked by the significance score.
}
\description{
rank single tf according to the rank of candidate TF pairs
}
