% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProScreenTF.R
\name{Screen_TF}
\alias{Screen_TF}
\title{screen switch-like TFs}
\usage{
Screen_TF(Switch_test, Tansient_test, top_n)
}
\arguments{
\item{lineage1}{refer to time series of committed data}
}
\value{
G1 & G2
}
\description{
function to fit time series data by nls
}
