% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_regression.R
\name{Switch_nls}
\alias{Switch_nls}
\title{screen switch-like and transient TFs}
\usage{
Switch_nls(scExp)
}
\arguments{
\item{lineage1}{refer to time series of committed data}
}
\value{
G1 & G2
}
\description{
function to fit time series data by nls
}
