% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two-Gene-BistableSwitch.R
\name{two_gene_bistableswitch_circuit_simulation}
\alias{two_gene_bistableswitch_circuit_simulation}
\title{calculate the Significance Score for candidate TF pairs}
\usage{
two_gene_bistableswitch_circuit_simulation(model_para, n_replicate)
}
\arguments{
\item{correction}{parameters to prevent the occurrence of zero entropy due to all computational data being zero, resulting in NA (Not Available) values}
}
\value{
A list of gene pair, ranked by the significance score.
}
\description{
function to fit time series data by nls
}
